fix performance

+++ components/tag/tag-card.tsx
@@ -5,26 +5,24 @@ import { TagName } from '@/components/tag/tag-name';
 import { cn } from '@/lib/utils';
 import Tag from '@/types/response/Tag';
 import TagTooltip from '@/components/tag/tag-tooltip';
+import { ErrorBoundary } from 'next/dist/client/components/error-boundary';
 
 type TagCardProps = HTMLAttributes<HTMLSpanElement> & {
   tag: Tag;
   onDelete?: (tag: Tag) => void;
 };
 
-export default function TagCard({
-  tag,
-  className,
-  onDelete,
-  ...props
-}: TagCardProps) {
+const TagCard = React.memo(_TagCard);
+
+export default TagCard;
+
+function _TagCard({ tag, className, onDelete, ...props }: TagCardProps) {
   const { name, value, color } = tag;
 
   const hasDelete = !!onDelete;
 
   const handleOnDelete = (tag: Tag) => {
-    if (onDelete) {
-      onDelete(tag);
-    }
+    if (onDelete) onDelete(tag);
   };
 
   return (
